program diagram
Activity diagram created in Microsoft Visio to help in the process of determining what the program will do in order to keep the project on the desired trajectory. This is also an opportunity to apply newly learned methods that will be helpful in developing and documenting future programs and program logic. And to just make me a better software developer in the long run. Because for me it helps to have the overall framework of what the program is supposed to do, step by step, to achieve a desired result. And, besides, look at my graph. I mean just look at it, what beaut.